+++ packages/server/migrations/0.11.3-non-nullable-asset-app-id.js
@@ -1,0 +1,29 @@
+import { DataTypes } from 'sequelize';
+
+export default {
+  key: '0.11.3',
+
+  async up(db) {
+    const queryInterface = db.getQueryInterface();
+    await queryInterface.changeColumn('Asset', 'AppId', {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: 'App',
+        key: 'id',
+      },
+    });
+  },
+
+  async down(db) {
+    const queryInterface = db.getQueryInterface();
+    await queryInterface.changeColumn('Asset', 'AppId', {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+      references: {
+        model: 'App',
+        key: 'id',
+      },
+    });
+  },
+};